Move apiv2 code to IBMQFactory
