+++ qiskit/providers/ibmq/ibmqprovider.py
@@ -19,17 +19,14 @@ from collections import OrderedDict
 
 from qiskit.providers import BaseProvider
 
-from .api_v2.clients import AuthClient, VersionClient
 from .credentials.configrc import remove_credentials
-from .credentials import (Credentials,
-                          read_credentials_from_qiskitrc, store_credentials, discover_credentials)
+from .credentials import (Credentials, read_credentials_from_qiskitrc,
+                          store_credentials, discover_credentials)
 from .exceptions import IBMQAccountError
-from .accountprovider import AccountProvider
 from .ibmqsingleprovider import IBMQSingleProvider
 
 
 QE_URL = 'https://quantumexperience.ng.bluemix.net/api'
-QX_AUTH_URL = 'https://auth.quantum-computing.ibm.com/api'
 
 
 class IBMQProvider(BaseProvider):
@@ -46,90 +43,6 @@ class IBMQProvider(BaseProvider):
         # keys are tuples (hub, group, project), as the convention is that
         # that tuple uniquely identifies a set of credentials.
         self._accounts = OrderedDict()
-
-        self._credentials = None
-        self._providers = OrderedDict()
-
-    def use_token(self, token, auth_url=QX_AUTH_URL, **kwargs):
-        """Authenticate against IBM Q Experience for use during this session.
-
-        Args:
-            token (str): IBM Q Experience API token.
-            auth_url (str): URL for the IBM Q Experience auth server.
-            **kwargs (dict): additional settings for the connection:
-                * proxies (dict): proxy configuration.
-                * verify (bool): verify the server's TLS certificate.
-
-        """
-        # TODO: rename function
-        # TODO: check and clean kwargs (verify str, proxies)
-        self._set_token(Credentials(token, auth_url, **kwargs))
-
-    def _set_token(self, credentials):
-        """Authenticate against IBM Q Experience and populate the providers.
-
-        Args:
-            credentials (Credentials): credentials for IBM Q Experience.
-
-        Raises:
-            IBMQAccountError:
-        """
-        # TODO: add checks (overwrite, mixing old and new)
-        version_finder = VersionClient(credentials.base_url)
-        version_info = version_finder.version()
-
-        if not (version_info['new_api'] and 'api-auth' in version_info):
-            raise IBMQAccountError(
-                'The URL specified ({}) is not a IBM Q Experience '
-                'authentication URL'.format(credentials.base_url))
-
-        auth_client = AuthClient(credentials.token,
-                                 credentials.base_url)
-
-        service_urls = auth_client.user_urls()
-        user_hubs = auth_client.user_hubs()
-
-        for hub_info in user_hubs:
-            # Build credentials.
-            provider_credentials = Credentials(
-                credentials.token,
-                url=service_urls['http'],
-                websockets_url=service_urls['ws'],
-                hub=hub_info['hub'],
-                group=hub_info['group'],
-                project=hub_info['project'],
-                proxies=credentials.proxies,
-                verify=credentials.verify)
-
-            # Build the provider.
-            provider = AccountProvider(provider_credentials,
-                                       auth_client.current_access_token())
-            self._providers[provider_credentials.unique_id()] = provider
-
-    def get_provider(self, hub=None, group=None, project=None):
-        """Return a provider for a single hub/group/project combination.
-
-        Returns:
-            AccountProvider:
-
-        Raises:
-            IBMQAccountError:
-        """
-        filters = []
-        for i, key in enumerate((hub, group, project)):
-            if key:
-                filters.append(lambda x: x[i] == key)
-
-        providers = list(filter(lambda x: all(f(x) for f in filters),
-                                self._providers.keys()))
-
-        if not providers:
-            raise IBMQAccountError('No providers matching the criteria')
-        if len(providers) > 1:
-            raise IBMQAccountError('More than one provider matching the '
-                                   'criteria')
-
-        return self._providers[providers[0]]
 
     def backends(self, name=None, filters=None, **kwargs):
         """Return all backends accessible via IBMQ provider, subject to optional filtering.
